requirements.txt
fastapi==0.110.0
uvicorn>=0.30.0
langchain-community>=0.4.0
chromadb>=1.5.0
langchain-chroma>=1.1.0
langchain-google-genai>=4.2.0
python-dotenv==1.0.1
pydantic>=2.10.1
PyYAML==6.0.1
mcp==1.2.0
langgraph>=1.0.8
langchain>=0.2.0
langserve[all]>=0.0.51
pydantic>=2.0
langgraph>=0.1.0
langchain-google-genai>=1.0.0
